+++ nori/graphviz_visualize.h
@@ -23,7 +23,7 @@ class GraphvizVisualizer {
                const nori::Morpheme* fromMorpheme, size_t toIndex,
                size_t toNodeId, const nori::Morpheme* toMorpheme,
                const std::string stringForm, int wordCost, int connectionCost,
-               size_t cost);
+               int cost);
   void addEos(size_t fromIndex, size_t fromNodeId,
               const nori::Morpheme* fromMorpheme);
   void finish();
